+++ NEWS.md
@@ -1,3 +1,8 @@
+# 2.1.1.9004
+
+* Added a new argument `likelihood.dist` to function `DefaultManyOpts` for user to
+specify the probability distribution to calculate the likelihood
+
 # 2.1.1.9003
 
 * Updated documentation for the return values of `DefaultManyOpts` to include three additional